Move #lap_splits_and_time_points into an Rspec support file and reference from rails_helper.rb.
