+++ spec/services/prior_split_time_finder_spec.rb
@@ -136,11 +136,4 @@ RSpec.describe PriorSplitTimeFinder do
       expect(finder.guaranteed_split_time.attributes).to eq(expected.attributes)
     end
   end
-
-  def lap_splits_and_time_points(event)
-    allow(event).to receive(:ordered_splits).and_return(event.splits)
-    lap_splits = event.required_lap_splits
-    time_points = lap_splits.map(&:time_points).flatten
-    [lap_splits, time_points]
-  end
 end
