+++ spec/services/segment_times_planner_spec.rb
@@ -172,11 +172,4 @@ RSpec.describe SegmentTimesPlanner do
       expect(times).to eq(expected)
     end
   end
-
-  def lap_splits_and_time_points(event)
-    allow(event).to receive(:ordered_splits).and_return(event.splits)
-    lap_splits = event.required_lap_splits
-    time_points = lap_splits.map(&:time_points).flatten
-    [lap_splits, time_points]
-  end
 end
