+++ spec/support/lap_split_methods.rb
@@ -1,0 +1,6 @@
+def lap_splits_and_time_points(event)
+  allow(event).to receive(:ordered_splits).and_return(event.splits)
+  lap_splits = event.required_lap_splits
+  time_points = lap_splits.map(&:time_points).flatten
+  [lap_splits, time_points]
+end
